test: move util tests
